Dinner time break, huge gains made

+++ cogs/queue.py
@@ -1,7 +1,8 @@
 # Author: Eddie Nieberding
-# Co-Authors: Gabby Khan
+# Co-Authors: Gabby Khan, Oliver Dininno
 import discord
 import queue
+from discord.utils import get
 from discord.ext import commands 
 
 
@@ -9,8 +10,9 @@ class queues(commands.Cog):
 
     def __init__(self, client):
         self.client = client
-        self.ohQueue = [] #Holds Student ID
-        self.ohMsg   = [] #Holds the message that was sent
+        self.ohQueue  = [] #Holds Student member
+        self.ohMsg    = [] #Holds the message that was sent
+        self.taOnDuty = []
     
     @commands.Cog.listener()
     async def on_ready(self):
@@ -19,40 +21,44 @@ class queues(commands.Cog):
     @commands.command(pass_context=True)
     async def join(self,ctx):
         " -Join the queue"
-        instructorID = ""
-        studMsg = "No message given"
-        if " " in ctx.message.content:
-            studMsg = " ".join(ctx.message.content.split()[1:])
-        
-        #Looks at all the channels and gets the id of the specific one
-        for channel in ctx.guild.channels:
-            if channel.name == "instructor-commands":
-                instructorID = channel.id
-            if channel.name == "request":
-                studnetID = channel.id
+        if len(self.taOnDuty) > 0:
+            instructorID = ""
+            studMsg = "No message given"
+            if " " in ctx.message.content:
+                studMsg = " ".join(ctx.message.content.split()[1:])
+            
+            #Looks at all the channels and gets the id of the specific one
+            for channel in ctx.guild.channels:
+                if channel.name == "instructor-commands":
+                    instructorID = channel.id
+                if channel.name == "request":
+                    studentID = channel.id
 
-        if ctx.message.channel.name == "request":
-            if ctx.author.id not in self.ohQueue:
-                studId = ctx.message.author.id
-                authID = "!reject "+str(studId)
+            if ctx.message.channel.name == "request":
+                if ctx.author not in self.ohQueue:
+                    studId = ctx.message.author
+                    authID = "!reject "+str(studId.id)
 
-                #Puts embedded message in the instructor channel
-                embedVar = discord.Embed(title=ctx.message.author.name, description=studMsg, color=0xff0000)
-                embedVar.add_field(name="To reject: ", value=authID, inline=False)
-                discordMsg = await ctx.guild.get_channel(instructorID).send(embed=embedVar)
+                    #Puts embedded message in the instructor channel
+                    embedVar = discord.Embed(title=ctx.message.author.display_name, description=studMsg, color=0xff0000)
+                    embedVar.add_field(name="To reject: ", value=authID, inline=False)
+                    discordMsg = await ctx.guild.get_channel(instructorID).send(embed=embedVar)
 
-                #adds the user and msg id to the queue
-                self.ohQueue.append(studId)
-                self.ohMsg.append(discordMsg)
+                    #adds the user and msg id to the queue
+                    self.ohQueue.append(studId)
+                    self.ohMsg.append(discordMsg)
+                else:
+                    msg = "{} You are already in the queue".format(ctx.message.author.mention)
+                    await ctx.send(msg,delete_after=3)
             else:
-                msg = "{} You are already in the queue".format(ctx.message.author.mention)
-                await ctx.send(msg,delete_after=3)
+                await ctx.send("You are in the wrong channel idiot.")
+            #deletes the command after being called 
+            await ctx.message.delete()
         else:
-            await ctx.send("You are in the wrong channel idiot.")
-        #deletes the command after being called 
-        await ctx.message.delete()
+            await ctx.send("The instructors are taking a break, please wait for them to return" )
       
     @commands.command(pass_context=True)
+    @commands.has_any_role('Professor','TA')
     async def reject(self,ctx):
         " -Leave the queue"
 
@@ -72,11 +78,9 @@ class queues(commands.Cog):
                 rejectionId = self.ohQueue.index(studentID)
 
                 #message returns user not other thing
-                print(self.ohQueue[rejectionId])
                 memId = self.ohQueue[rejectionId]
                 guildId = await ctx.guild.fetch_member(studentID)
                 await guildId.send(studMsg)
-                print(guildId)
 
                 await self.ohMsg[rejectionId].delete()
                 await ctx.message.delete()
@@ -114,14 +118,14 @@ class queues(commands.Cog):
             index = -1
         if index != -1:
             # finds the student ID, gives a message and finsd rejectionID
-            studnetID = ctx.message.author.id
+            studentID = ctx.message.author.id
             studMsg = "You have decided to leave the queue!"
-            rejectionId = self.ohQueue.index(studnetID)
+            rejectionId = self.ohQueue.index(studentID)
 
             #message returns user not other thing
             print(self.ohQueue[rejectionId])
             memId = self.ohQueue[rejectionId]
-            guildId = await ctx.guild.fetch_member(studnetID)
+            guildId = await ctx.guild.fetch_member(studentID)
             await guildId.send(studMsg)
             print(guildId)
 
@@ -130,6 +134,107 @@ class queues(commands.Cog):
             self.ohMsg.pop(rejectionId)
             self.ohQueue.pop(rejectionId)
 
+    @commands.command(pass_context=True)
+    @commands.has_any_role('Professor','TA')
+    async def startOh(self,ctx):
+        ta = ctx.message.author
+        if ta.id not in self.taOnDuty:
+            if len(self.taOnDuty) == 0:
+                channel = get(ta.guild.channels,name="request")
+                embedVar = discord.Embed(title="Office hours are open!", description="Use !join [reason] to join", color=0x00ff00)
+                await channel.send(embed=embedVar)    
+            self.taOnDuty.append(ta.id)
+            await ctx.send("You have checking into office hours!",delete_after=5)
+        else:
+            await ctx.send("You are already in the office hours")
+            
+
+    @commands.command(pass_context=True)
+    @commands.has_any_role('Professor','TA')
+    async def endOh(self,ctx):
+        ta = ctx.message.author
+        if ta.id in self.taOnDuty:
+           self.taOnDuty.remove(ta.id)
+           await ctx.send("You have checked out of office hours!", delete_after=5)
+           if len(self.taOnDuty) == 0:
+                channel = get(ta.guild.channels, name="request")
+                embedVar = discord.Embed(title="Office hours are now over!", description="Please wait for a TA to open Office hours to join", color=0xff0000)
+                await channel.send(embed=embedVar)
+                #tells all students in the queue they have been removed
+                for x in range(len(self.ohMsg)):
+                    await self.ohMsg[x].delete()
+                    await self.ohQueue[x].send("Office Hours are now closed. You have been removed from the queue.")
+                self.ohMsg.clear()
+                self.ohQueue.clear()
+                    
+                
+        else:
+            await ctx.send("You are not in office hours")
+    
+    @commands.command(pass_context=True)
+    @commands.has_any_role('Professor','TA')
+    async def accept(self,ctx):
+        ta = ctx.message.author
+        if ta.id in self.taOnDuty:
+            message = self.ohMsg.pop(0)
+            studentMember = self.ohQueue.pop(0)
+            
+
+            member = ctx.message.author
+            myRoles = ctx.guild.roles #[Everyone, Limbo, Student, TA, Professor, HACKER!!, BOT]
+
+            #creates objects of each role for easy access
+            profRole = get(member.guild.roles, name="Professor")
+            taRole = get(member.guild.roles, name="TA")
+            stuRole = get(member.guild.roles, name="Student")
+            everyRole = get(member.guild.roles, name="@everyone")
+            #create a temp role to assign to student
+            name = studentMember.display_name
+            await ctx.guild.create_role(name=name)
+            tempRole = get(member.guild.roles, name=studentMember.display_name)
+            await studentMember.add_roles(tempRole)
+            
+            newCategory = await ctx.guild.create_category("{} Office Hour Session".format(studentMember.display_name))
+            await newCategory.set_permissions(everyRole,read_messages=False)
+            await newCategory.set_permissions(stuRole,read_messages=False)
+            await newCategory.set_permissions(tempRole, read_messages=True)
+            await newCategory.set_permissions(taRole, read_messages=True)
+            await newCategory.set_permissions(profRole, read_messages=True)
+
+            TC = await newCategory.create_text_channel("{} Text Chat".format(studentMember.display_name))
+            VC = await newCategory.create_voice_channel("Session Voice Chat".format(studentMember.display_name))
+            await TC.send("{} and {}, time to figure out those bugs! Use !close to end the session.".format(studentMember.mention,ta.mention ))
+            
+
+        else:
+            await ctx.send("You are not in office hours")
+
+    @commands.command(pass_context=True)
+    async def close(self,ctx):
+
+        untouchable = ["Instructors","Students","Authentication"]
+        message = ctx.message
+        textChannel = message.channel
+        categoryId = textChannel.category_id
+        category = get(textChannel.guild.categories, id = categoryId)
+        #deletes a role if called from inside of office hours
+        if "Office Hour Session" in category.name:
+            catName = category.name
+            userName = catName.split(" Office",1)
+            tempRole = get(message.guild.roles, name = userName[0])
+            await tempRole.delete()
+        #deletes all items under a category
+        if category.name not in untouchable:
+            for vc in category.voice_channels:
+                await vc.delete()
+            for tc in category.text_channels:
+                await tc.delete()
+            await category.delete()
+        else:
+            await ctx.send("You are not in the correct chat to call this command.")
+        
+
+        
 
     @commands.command(pass_context=True)
     async def request_help(self,ctx):
